refac general: zeros_like and full
