+++ ivy/functional/backends/numpy/creation.py
@@ -127,6 +127,22 @@ def arange(stop, start=0, step=1, dtype=None, dev=None):
             return res.astype(np.int32)
     return res
 
+
+
+# noinspection PyShadowingNames
+def zeros_like(x, dtype=None, dev=None):
+    if dtype:
+        dtype = 'bool_' if dtype == 'bool' else dtype
+        dtype = np.__dict__[dtype]
+    else:
+        dtype = x.dtype
+    return _to_dev(np.zeros_like(x, dtype=dtype), dev)
+
+
+def full(shape, fill_value, dtype=None, device=None):
+    return _to_dev(np.full(shape, fill_value, dtype_from_str(default_dtype(dtype, fill_value))), device)
+
+
 # Extra #
 # ------#
