+++ ivy/functional/backends/tensorflow/old/general.py
@@ -100,18 +100,6 @@ astype = cast
 
 
 
-# noinspection PyShadowingNames
-def zeros_like(x, dtype=None, dev=None):
-    dtype = tf.__dict__[dtype] if dtype else dtype
-    dev = default_device(dev)
-    with tf.device(dev_from_str(dev)):
-        return tf.zeros_like(x, dtype=dtype)
-
-
-def full(shape, fill_value, dtype=None, device=None):
-    with tf.device(dev_from_str(default_device(device))):
-        return tf.fill(shape, tf.constant(fill_value, dtype=dtype_from_str(default_dtype(dtype, fill_value))))
-
 
 def one_hot(indices, depth, dev=None):
     dev = default_device(dev)
